Separate API helpers test cases into separate files
